refactor(pkey): split src/pkey.c as multi modules in src/pkey
Others:

- macOS static build with -Wl,-all_load for libopenssl.a
- Add OPENSSL_STATIC support for macOS using -Wl,-all_load flag
- Set MACOSX_DEPLOYMENT_TARGET=10.12 for compatibility
- Fix linker error when building with static OpenSSL on macOS
